pyproject.toml
[project]
name = "attendance-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "cmake>=4.2.3",
    "dlib>=20.0.0",
    "face-recognition>=1.3.0",
    "face-recognition-models",
    "numpy>=2.0.2",
    "opencv-python>=4.13.0.92",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "setuptools<71",
]

[tool.uv.sources]
face-recognition-models = { git = "https://github.com/ageitgey/face_recognition_models" }
